docs: Documentation of Next.js multizone apps with authentication sharing and refactored routing
