+++ bciers/apps/registration/project.json
@@ -66,7 +66,7 @@
     "e2e:sequential": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "playwright test --workers=1 client/e2e/*"
+        "command": "playwright test --workers=1 bciers/apps/registration1/e2e/*"
       }
     }
   }
