+++ bciers/apps/registration1/app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,2 @@
-import { handlers } from "@/auth";
+import { handlers } from "@/dashboard/auth";
 export const { GET, POST } = handlers;
