+++ bciers/apps/registration1/e2e/pages/onboarding/home/page.spec.ts
@@ -1,4 +1,4 @@
-// 🧪 Suite to test `client/app/(onboarding)/home/page.tsx`
+// 🧪 Suite to test `bciers/apps/registration1/app/(onboarding)/home/page.tsx`
 import { test, expect } from "@playwright/test";
 // 🪄 Page Object Models
 import { DashboardPOM } from "@/e2e/poms/dashboard";
