+++ bciers/apps/registration1/project.json
@@ -60,7 +60,7 @@
     "e2e:sequential": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "playwright test --workers=1 client/e2e/*"
+        "command": "playwright test --workers=1 bciers/apps/registration1/e2e/*"
       }
     }
   }
