+++ docs/README.md
@@ -1,9 +1,11 @@
 # Documentation
 
-| Name                        | Link                                                                                                                                                                                                                                                                                                           |
-| --------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| Developer Environment Setup | [developer-environment-setup.md](./developer-environment-setup.md)                                                                                                                                                                                                                                             |
-| CI                          | [ci.md](./ci.md)                                                                                                                                                                                                                                                                                               |
-| Backend                     | [api-design.md](./backend/api-design.md) - [data-models.md](./backend/data-models.md) - [entity-relationship-diagram.md](./backend/entity-relationship-diagram.md) - [misc.md](./backend/misc.md) - [mypy.md](./backend/mypy.md) - [optimization.md](./backend/optimization.md) - [test.md](./backend/test.md) |
-| Devops                      | [helm-secrets-cas-pipeline.md](./devops/helm-secrets-cas-pipeline.md) - [release.md](./devops/release.md) - [state-and-container-setup.md](./devops/state-and-container-setup.md)                                                                                                                              |
-| Frontend                    | [css-theme-guide.md](./frontend/css-theme-guide.md) - [developer-guide.md](./frontend/developer-guide.md) - [growthbook.md](./frontend/growthbook.md) - [nx-monorepo.md](./frontend/nx-monorepo.md) - [rjsf-forms.md](./frontend/rjsf-forms.md)                                                                |
+| Name                        | Link                                                                                                                                                                                                                                                                                                                      |
+| --------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| Developer Environment Setup | [developer-environment-setup.md](./developer-environment-setup.md)                                                                                                                                                                                                                                                        |
+| CI                          | [ci.md](./ci.md)                                                                                                                                                                                                                                                                                                          |
+| Backend                     | [api-design.md](./backend/api-design.md) - [data-models.md](./backend/data-models.md) - [entity-relationship-diagram.md](./backend/entity-relationship-diagram.md) - [misc.md](./backend/misc.md) - [mypy.md](./backend/mypy.md) - [optimization.md](./backend/optimization.md) - [test.md](./backend/test.md)            |
+| Devops                      | [helm-secrets-cas-pipeline.md](./devops/helm-secrets-cas-pipeline.md) - [release.md](./devops/release.md) - [state-and-container-setup.md](./devops/state-and-container-setup.md)                                                                                                                                         |
+| Frontend                    | [css-theme-guide.md](./frontend/css-theme-guide.md) - [developer-guide.md](./frontend/developer-guide.md) - [growthbook.md](./frontend/growthbook.md) - [nx-monorepo.md](./frontend/nx-monorepo.md) - [rjsf-form-base-components.md](./frontend/rjsf-form-base-components.md) - [rjsf-forms.md](./frontend/rjsf-forms.md) |
+| Frontend\Testing            | [playwright.md](./frontend/testing/playwright.md) - [vitest.md](./frontend/testing/vitest.md)                                                                                                                                                                                                                             |
+| Registration Part 2         | [backend-structure.md](./registration_part_2/backend-structure.md) - [data-models.md](./registration_part_2/data-models.md) - [frontend-development.md](./registration_part_2/frontend-development.md)                                                                                                                    |
